# [SDK-453] Feature/precompile support