--- a/Runtime/AvatarManager.cs
+++ b/Runtime/AvatarManager.cs
@@ -43,7 +43,7 @@ namespace ReadyPlayerMe.AvatarCreator
             inCreatorAvatarLoader = new InCreatorAvatarLoader();
             avatarAPIRequests = new AvatarAPIRequests(ctxSource.Token);
         }
-        
+
         /// <summary>
         /// Create a new avatar from a provided template.
         /// </summary>
@@ -65,7 +65,7 @@ namespace ReadyPlayerMe.AvatarCreator
                 OnError?.Invoke(e.Message);
                 return avatarProperties;
             }
-            
+
             if (ctxSource.IsCancellationRequested)
             {
                 return avatarProperties;
@@ -119,6 +119,25 @@ namespace ReadyPlayerMe.AvatarCreator
             }
 
             return await inCreatorAvatarLoader.Load(avatarId, bodyType, gender, data);
+        }
+
+        public async void PrecompileAvatar(string avatarId, PrecompileData precompileData)
+        {
+            try
+            {
+                avatarAPIRequests.PrecompileAvatar(avatarId, precompileData, avatarConfigParameters);
+                Debug.Log("Precompiled avatar.");
+            }
+            catch (Exception e)
+            {
+                OnError?.Invoke(e.Message);
+                Debug.LogError("Precompiled avatar request failed.");
+            }
+
+            if (ctxSource.IsCancellationRequested)
+            {
+                Debug.LogWarning("Precompiled avatar request cancelled.");
+            }
         }
 
         /// <summary>
@@ -199,7 +218,7 @@ namespace ReadyPlayerMe.AvatarCreator
 
             return avatarId;
         }
-        
+
         /// <summary>
         /// This will delete the avatar draft which have not been saved. 
         /// </summary>
@@ -229,7 +248,7 @@ namespace ReadyPlayerMe.AvatarCreator
                 OnError?.Invoke(e.Message);
             }
         }
-        
+
         public async Task<ColorPalette[]> LoadAvatarColors()
         {
             ColorPalette[] colors = null;