--- a/Runtime/Data/PrecompileData.cs
+++ b/Runtime/Data/PrecompileData.cs
@@ -1,0 +1,11 @@
+using System;
+using System.Collections.Generic;
+
+namespace ReadyPlayerMe.AvatarCreator
+{
+    [Serializable]
+    public struct PrecompileData
+    {
+        public Dictionary<string, string[]> data;
+    }
+}