--- a/Runtime/Data/PrecompileData.cs.meta
+++ b/Runtime/Data/PrecompileData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5f3c3e48f9bcda14d8f21a84ac630dca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: