--- a/Runtime/JsonHelpers/AvatarPropertiesExtensions.cs
+++ b/Runtime/JsonHelpers/AvatarPropertiesExtensions.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 
@@ -29,38 +28,6 @@ namespace ReadyPlayerMe.AvatarCreator
                 settings.NullValueHandling = NullValueHandling.Ignore;
                 settings.DefaultValueHandling = DefaultValueHandling.Ignore;
             }
-
-            return JsonConvert.SerializeObject(data, settings);
-        }
-
-        public static string GetPrecompilePayload(this AvatarProperties avatarProperties)
-        {
-            var contractResolver = new DefaultContractResolver
-            {
-                NamingStrategy = new CamelCaseNamingStrategy()
-            };
-            var settings = new JsonSerializerSettings
-            {
-                ContractResolver = contractResolver,
-                Formatting = Formatting.Indented
-            };
-
-            var dataContainer = new Dictionary<string, string[]>();
-            foreach (var keyValuePair in avatarProperties.Assets)
-            {
-                if (string.IsNullOrEmpty(keyValuePair.Value.ToString())) continue;
-                var key = keyValuePair.Key.ToString();
-                var camelCaseKey = $"{char.ToLowerInvariant(key[0])}{key.Substring(1)}";
-                dataContainer.Add(camelCaseKey, new[] { keyValuePair.Value.ToString() });
-            }
-
-            var data = new Dictionary<string, Dictionary<string, string[]>>
-            {
-                { "data", dataContainer }
-            };
-
-            settings.NullValueHandling = NullValueHandling.Ignore;
-            settings.DefaultValueHandling = DefaultValueHandling.Ignore;
 
             return JsonConvert.SerializeObject(data, settings);
         }