--- a/Runtime/Utils/StringExtension.cs.meta
+++ b/Runtime/Utils/StringExtension.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9300f547d88e0da40908df2f775dbbdf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: