--- a/Samples~.meta
+++ b/Samples~.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0a2b5011a601ea744a008b5adb3f2c62
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: