--- a/Samples~/Scripts/UI/SelectionScreens/AvatarCreatorSelection.cs
+++ b/Samples~/Scripts/UI/SelectionScreens/AvatarCreatorSelection.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using ReadyPlayerMe.AvatarCreator;
@@ -98,7 +99,7 @@ namespace ReadyPlayerMe
                 categoryUICreator.SetActiveCategoryButtons(true);
             }
         }
-        
+
         private void Cleanup()
         {
             if (currentAvatar != null)
@@ -115,14 +116,14 @@ namespace ReadyPlayerMe
 
         private void OnErrorCallback(string error)
         {
-            SDKLogger.Log(TAG,$"An error occured: {error}");
+            SDKLogger.Log(TAG, $"An error occured: {error}");
             avatarManager.OnError -= OnErrorCallback;
             partnerAssetManager.OnError -= OnErrorCallback;
 
             ctxSource?.Cancel();
             StateMachine.GoToPreviousState();
             LoadingManager.EnableLoading(error, LoadingManager.LoadingType.Popup, false);
-            SDKLogger.Log(TAG,"Going to previous state");
+            SDKLogger.Log(TAG, "Going to previous state");
         }
 
         private async Task LoadAssets()
@@ -137,7 +138,7 @@ namespace ReadyPlayerMe
 
             CreateUI(AvatarCreatorData.AvatarProperties.BodyType);
             categoriesAssetsLoaded = new List<Category>();
-            
+
             await partnerAssetManager.GetAssets();
             await CreateAssetsByCategory(Category.FaceShape);
 
@@ -147,6 +148,7 @@ namespace ReadyPlayerMe
         private async void OnCategorySelected(Category category)
         {
             await CreateAssetsByCategory(category);
+            avatarManager.PrecompileAvatar(AvatarCreatorData.AvatarProperties.Id, partnerAssetManager.GetPrecompileData(new[] { category }));
         }
 
         private async Task<GameObject> LoadAvatar()
@@ -202,7 +204,6 @@ namespace ReadyPlayerMe
             saveButton.gameObject.SetActive(true);
         }
 
-
         private async Task CreateAssetsByCategory(Category category)
         {
             if (!categoriesAssetsLoaded.Contains(category))
@@ -215,6 +216,8 @@ namespace ReadyPlayerMe
             }
 
             var assets = partnerAssetManager.GetAssetsByCategory(category);
+
+            var firstTenArray = assets.Take(10).ToArray();
             if (assets == null || assets.Count == 0)
             {
                 return;