<!-- Copy the TICKETID for this task from Monday and add it to the PR name in brackets -->
<!-- PR name should look like: [TICKETID] My Pull Request -->

<!-- Add link for the ticket here editing the TICKETID-->

## [SDK-453](https://ready-player-me.atlassian.net/browse/SDK-453)

## Description

-  this adds support for using the precompile endpoint to slightly increase avatar update/load times during avatar creation process.

<!-- Fill the section below with Added, Updated and Removed information. -->
<!-- If there is no item under one of the lists remove it's title. -->


<!-- Testability -->

## How to Test

- Import CAC package and run the sample scene. 
- switch categories  and wait  a few seconds (as precompile takes some time to complete
- then switch to one of the first 20 assets of that category
- In theory, these first 20 should generally load faster (0.5s -> 0.8s for outfits) compared to others (~1s -> 2s)

<!-- Update your progress with the task here -->

## Checklist

-   [ ] Tests written or updated for the changes.
-   [ ] Documentation is updated.
-   [ ] Changelog is updated.

<!--- Remember to copy the Changes Section into the commit message when you close the PR -->





[SDK-453]: https://ready-player-me.atlassian.net/browse/SDK-453?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ